정책 관련 rule, preference 적용

+++ src/main/kotlin/com/the/playting/policy/service/PolicyService.kt
@@ -1,7 +1,7 @@
 package com.the.playting.policy.service
 import com.the.playting.policy.model.dto.PolicyRequestDto
 import com.the.playting.policy.model.dto.PolicyResponseDto
-import com.the.playting.policy.repository.PolicyRepository
+import com.the.playting.policy.infrastructure.repository.PolicyRepository
 import org.springframework.data.domain.Sort
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
@@ -13,7 +13,7 @@ class PolicyService(private val policyRepository: PolicyRepository) {
         return policyRepository.findAll(Sort.by(Sort.Order.asc("createdAt"))).map { it.toDto() }
     }
 
-    fun getPolicyById(id: Int): PolicyResponseDto? {
+    fun getPolicyById(id: Long): PolicyResponseDto? {
         return policyRepository.findByIdOrNull(id)?.toDto();
     }
 
@@ -22,13 +22,13 @@ class PolicyService(private val policyRepository: PolicyRepository) {
         return policyRepository.save(policy).toDto()
     }
 
-    fun updatePolicy(id: Int, policyDto: PolicyRequestDto): PolicyResponseDto? {
+    fun updatePolicy(id: Long, policyDto: PolicyRequestDto): PolicyResponseDto? {
         val existingPolicy = policyRepository.findByIdOrNull(id) ?: throw IllegalArgumentException("Policy not found")
 
         return policyRepository.save(existingPolicy).toDto()
     }
 
-    fun deletePolicy(id: Int) {
+    fun deletePolicy(id: Long) {
         policyRepository.deleteById(id)
     }
